+++ packages/formula-app-e2e/src/support/app.po.ts
@@ -1,1 +1,3 @@
 export const getGreeting = () => cy.get('h1');
+
+export const getLoginErrors = () => cy.get('.errors');